+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Bugfixes
 
+- `no-implicit-close` now provides more context when closed by a sibling.
 - `close-order` no longer reports error for implicitly closed elements.
 
 ## 0.11.0 (2018-09-23)